[workflow] Add start-performance-comparison.yml
We intend to run performance comparisons on a daily basis. This new
workflow does the first step by running every night close after
midnight.

As a simplification we assume that after 24 hours we have all the
`pgo` and `big-merge` COPR builds ready from the day before.

The new `start-perf-comparison` of the the `main.py` program will be
added in a later PR. To make it easy to follow the PRs I decided to
bring in the necessary changes piece by piece.